==> 02-bind.Rmd <==
---
title: "bind"
author: "Elise Boos"
date: "2022-10-31"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r read in}
require(tidyverse)

data2014 <- read.csv("data2014_final.csv", stringsAsFactors = T)
data2015 <- read.csv("data2015_final.csv", stringsAsFactors = T)
data2016 <- read.csv("data2016_final.csv", stringsAsFactors = T)
data2017 <- read.csv("data2017_final.csv", stringsAsFactors = T)
data2018 <- read.csv("data2018_final.csv", stringsAsFactors = T)
data2019 <- read.csv("data2019_final.csv", stringsAsFactors = T)
data2020 <- read.csv("data2020_final.csv", stringsAsFactors = T)

camop_2014 <- read.csv("camop_2014.csv", stringsAsFactors = T)
camop_2015 <- read.csv("camop_2015.csv", stringsAsFactors = T)
camop_2016 <- read.csv("camop_2016.csv", stringsAsFactors = T)
camop_2017 <- read.csv("camop_2017.csv", stringsAsFactors = T)
camop_2018 <- read.csv("camop_2018.csv", stringsAsFactors = T)
camop_2019 <- read.csv("camop_2019.csv", stringsAsFactors = T)
camop_2020 <- read.csv("camop_2020.csv", stringsAsFactors = T)

sitecovs2014 <- read.csv("sitecovs2014.csv", stringsAsFactors = T)
sitecovs2015 <- read.csv("sitecovs2015.csv", stringsAsFactors = T)
sitecovs2016 <- read.csv("sitecovs2016.csv", stringsAsFactors = T)
sitecovs2017 <- read.csv("sitecovs2017.csv", stringsAsFactors = T)
sitecovs2018 <- read.csv("sitecovs2018.csv", stringsAsFactors = T)
sitecovs2019 <- read.csv("sitecovs2019.csv", stringsAsFactors = T)
sitecovs2020 <- read.csv("sitecovs2020.csv", stringsAsFactors = T)

```

cam op dataframe: site, easting, northing, Date.Placement, Last.Record, Days
```{r bind cams}
colnames(camop_2014) <- c("site", "Date.Placement", "Days", "Last.record", "Northing", "Easting")
colnames(camop_2015) <- c("site", "Date.Placement", "Days", "Last.record", "Northing", "Easting")
colnames(camop_2016) <- c("site", "Date.Placement", "Days", "Last.record", "Northing", "Easting")
colnames(camop_2017) <- c("site", "Date.Placement", "Days", "Last.record", "Northing", "Easting", "Logging")
camop_2017 <- select(camop_2017, 1:6)
colnames(camop_2018) <- c("site", "Date.Placement", "Days", "Last.record", "Northing", "Easting")
colnames(camop_2019) <- c("site", "Date.Placement", "Days", "Last.record", "Northing", "Easting")
colnames(camop_2020) <- c("site", "Date.Placement", "Days", "Last.record", "Northing", "Easting")

allcamop <- rbind(camop_2014, camop_2015, camop_2016, camop_2017, camop_2018, camop_2019, camop_2020)

#write.csv(allcamop, "allcamop.csv", row.names = F)

camop <- allcamop

#save to source
save(camop, file = "camop.RData")
```



```{r bind data}

data2014_select <- select(data2014, site, Picture, Species, Nr, Year, Month, Day, Hour_exact, obsdate, Mammal, Easting.x, Northing.x, date_time_obs)

data2015_select <- select(data2015, site, Picture, Species, Nr, Year, Month, Day, Hour_exact, obsdate, Mammal, Easting.1, Northing.1, date_time_obs)

data2016_select <- select(data2016, site, picture, species.name, Nr, Year, Month, Day, Hour.Exact, obsdate, Mammal, Easting, Northing, date_time_obs)

#add mammal column to 2017
data2017$species.name <- gsub("Gray fox", "Gray Fox", data2017$species.name)
data2017$species.name <- gsub("Ruddy Quail Dove ", "Ruddy Quail-dove", data2017$species.name)
data2017$species.name <- gsub("Collared Pecary", "Collared Peccary", data2017$species.name)
data2017$species.name <- gsub("Crested guan", "Crested Guan", data2017$species.name)
data2017$species.name <- gsub("Lesson's motmot", "Lesson's Motmot", data2017$species.name)
data2017$species.name <- gsub("Bare-throated Tiger-heron", "Bare-throated Tiger Heron", data2017$species.name)
data2017$species.name <- gsub("Black-crowned Night-Heron", "Black-crowned Night Heron", data2017$species.name)
data2017$species.name <- gsub("Great Blue-heron", "Great Blue Heron", data2017$species.name)
data2017$species.name <- gsub("Clay coloured Thrush", "Clay-coloured Thrush", data2017$species.name)
data2017$species.name <- gsub("Yellow-crowned Night-Heron", "Yellow-crowned Night Heron", data2017$species.name)
data2017$species.name <- gsub("Blue Ground-Dove", "Blue Ground Dove", data2017$species.name)
data2017$species.name <- gsub("Gray catbird", "Gray Catbird", data2017$species.name)

#add mammal Y/N
data2017$Mammal <- ifelse(data2017$species.name %in% c("Baird's Tapir", "Puma", "Red Brocket Deer", "White-lipped Peccary", 
                                                               "White-tailed Deer", "Ocelot", "Agouti", "Gray Fox", 
                                                               "Tamandua", "White-nosed Coati", "Common Opossum", "Jaguar", 
                                                               "Margay", "Paca", "Tayra", "Racoon", "Striped Hog-nosed Skunk", 
                                                               "Otter", "Brown Four-eyed Opossum", "Nine-banded Armadillo", 
                                                               "Coyote", "Jaguarundi", "Rodent", "Social Flycatcher", "Squirrel", 
                                                               "Deppe's Squirrel"), "Yes",
                              ifelse(data2017$species.name %in% c("Great Curassow", "Ocellated Turkey", "Unknown", "Humans", 
                                                                      "Mottled Owl", "Neeltje", "Gray-headed Dove", "Chachalaca", 
                                                                      "Ruddy Quail-dove", "Black-faced Antthrush", "Scaled Pigeon", 
                                                                      "Collared Peccary", "Lesson's Motmot", "Wood Thrush", 
                                                                      "Boat-billed Heron", "Slaty-breasted Tinamou", "Bare-throated Tiger Heron", 
                                                                      "Black-crowned Night Heron", "Cane Toad", "Great Blue Heron", 
                                                                      "Great Tinamou", "Limpkin", "Agami Heron", "Clay-coloured Thrush", 
                                                                      "Yellow-crowned Night Heron", "Russet-naped Woodrail", "Bird", 
                                                                      "Blue Ground Dove", "Gray Catbird", "Gray-chested Dove", "Hawk", 
                                                                      "Northern Waterthrush", "Ornate Hawk-eagle", "Roadside Hawk", 
                                                                      "Little Tinamou", "Crested Guan"), "No", NA))


data2017_select <- select(data2017, site, picture, species.name, Nr, Year, Month, Day, Hour.Exact, obsdate, Mammal, Easting, Northing, date_time_obs)

data2018_select <- select(data2018, site, picture, species.name, Nr, Year, Month, Day, Hour.Exact, obsdate, Mammal, Easting.y, Northing.y, date_time_obs)

data2019_select <- select(data2019, site, picture, species.name, Nr, Year, Month, Day, Hour.Exact, obsdate, Mammal, Easting.y, Northing.y, date_time_obs)

data2020_select <- select(data2020, Name, picture, species.name, Nr, Year, Month, Day, Hour.Exact, obsdate, Mammal, Easting.x, Northing.x, date_time_obs)

#change col names before binding
colnames(data2014_select) <- c("site", "picture", "Species", "Number_recorded", "Year", "Month", "Day", "Hour", "obsdate", "Mammal", "Easting", "Northing", "date_time_obs")
colnames(data2015_select) <- c("site", "picture", "Species", "Number_recorded", "Year", "Month", "Day", "Hour", "obsdate", "Mammal", "Easting", "Northing", "date_time_obs")
colnames(data2016_select) <- c("site", "picture", "Species", "Number_recorded", "Year", "Month", "Day", "Hour", "obsdate", "Mammal", "Easting", "Northing", "date_time_obs")
colnames(data2017_select) <- c("site", "picture", "Species", "Number_recorded", "Year", "Month", "Day", "Hour", "obsdate", "Mammal", "Easting", "Northing", "date_time_obs")
colnames(data2018_select) <- c("site", "picture", "Species", "Number_recorded", "Year", "Month", "Day", "Hour", "obsdate", "Mammal", "Easting", "Northing", "date_time_obs")
colnames(data2019_select) <- c("site", "picture", "Species", "Number_recorded", "Year", "Month", "Day", "Hour", "obsdate", "Mammal", "Easting", "Northing", "date_time_obs")
colnames(data2020_select) <- c("site", "picture", "Species", "Number_recorded", "Year", "Month", "Day", "Hour", "obsdate", "Mammal", "Easting", "Northing", "date_time_obs")

#bind
dat <- rbind(data2014_select, data2015_select, data2016_select, data2017_select, data2018_select, data2019_select, data2020_select)

#save
save(dat, file = "dat.RData")
```



```{r bind site covs}

#add logging column? if ever logged?
#sitecov2014$Logged.in.Year <- replace_na(sitecov2014$Logged.in.Year, "No") #na to No
#sitecov2014$Year_Since_Logging <- ifelse(sitecov2014$Logged.in.Year == "No", "No","Yes")



#add season column
#2014
sitecovs2014$Month <- str_replace(sitecovs2014$Date.Placement.x, pattern = ".*-([^-]*)-.*", replacement = "\\1")
sitecovs2014$Season <- ifelse(sitecovs2014$Month %in% c("05", "06", "07", "08", "09", "10"), 
                             "Rainy", "Dry")
#2015
sitecovs2015$Month <- str_replace(sitecovs2015$Date_Place.x, pattern = ".*-([^-]*)-.*", replacement = "\\1")
sitecovs2015$Season <- ifelse(sitecovs2015$Month %in% c("05", "06", "07", "08", "09", "10"), 
                             "Rainy", "Dry")
#2016
sitecovs2016$Month <- str_replace(sitecovs2016$Date.Placement.x, pattern = ".*-([^-]*)-.*", replacement = "\\1")
sitecovs2016$Season <- ifelse(sitecovs2016$Month %in% c("05", "06", "07", "08", "09", "10"), 
                             "Rainy", "Dry")
#2017
sitecovs2017$Month <- str_replace(sitecovs2017$Date.Placement, pattern = ".*-([^-]*)-.*", replacement = "\\1")
sitecovs2017$Season <- ifelse(sitecovs2017$Month %in% c("05", "06", "07", "08", "09", "10"), 
                               "Rainy", "Dry")
#2018
sitecovs2018$Month <- str_replace(sitecovs2018$Date.Placement, pattern = ".*-([^-]*)-.*", replacement = "\\1")
sitecovs2018$Season <- ifelse(sitecovs2018$Month %in% c("05", "06", "07", "08", "09", "10"), 
                             "Rainy", "Dry")
sitecovs2018$Active_Log <- ifelse(sitecovs2018$Active_Log == "no", "No", "Yes") #fix error
#2019
sitecovs2019$Month <- str_replace(sitecovs2019$Date.Placement.x, pattern = ".*-([^-]*)-.*", replacement = "\\1")
sitecovs2019$Season <- ifelse(sitecovs2019$Month %in% c("05", "06", "07", "08", "09", "10"), 
                             "Rainy", "Dry")
#2020
sitecovs2020$Month <- str_replace(sitecovs2020$Date.Placement.x, pattern = ".*-([^-]*)-.*", replacement = "\\1")
sitecovs2020$Season <- ifelse(sitecovs2020$Month %in% c("05", "06", "07", "08", "09", "10"), 
                             "Rainy", "Dry")

#select columns
sitecovs2014 <- select(sitecovs2014, site, Brand.x, Conservation.Area, Active.Logging, Hurricane.Damage, Fire.Damage, Location.descr., Landscape, Lat, Lon, Season)
#get location???/ property 

sitecovs2015 <- select(sitecovs2015, site, Brand, Conservati, Active_Log, Hurricane, Fire_Damag, Location_d, Landscape, Lat, Lat.1, Season)
#Lat.1 is Long

sitecovs2016 <- select(sitecovs2016, site, Brand, Conservati, Active_Log, Hurricane, Fire_Damag, Location_d, Landscape, Lat, Long, Season)
#this one has property ----- probs should add in 

#fill data at some point- missing from sitecovs2017
#sitecovs2017 <- sitecovs2017$Brand %>% fill(sitecovs2017$Brand, .direction = "down")

sitecovs2017 <- select(sitecovs2017, site, Brand,  Conservati, Active_Log, Hurricane, Fire_Damag, Location_d, Landscape, Lat, Long, Season)

#missing some info
sitecovs2018$Conservation <- NA
sitecovs2018 <- select(sitecovs2018, site, Brand, Conservation, Active_Log, Hurricane, Fire_Damag, Location_d, Landscape, Lat, Long, Season)

#missing things
sitecovs2019$Conservation <- NA
sitecovs2019$Hurricane <- NA
sitecovs2019$Fire <- NA
sitecovs2019$Location <- NA
sitecovs2019$Landscape <- NA
sitecovs2019 <- select(sitecovs2019, site, Brand, Conservation, Logging, Hurricane, Fire, Location, Landscape, Lat, Long, Season)

sitecovs2020$Conservation <- NA
sitecovs2020$Fire <- NA
sitecovs2020$Hurricane <- NA
sitecovs2020$Location <- NA
sitecovs2020$Landscape <- NA
sitecovs2020 <- select(sitecovs2020, Name, Brand, Conservation, Logging, Hurricane, Fire, Location, Landscape, Lat, Long, Season)

#colnames
colnames(sitecovs2014) <- c("site", "Brand", "Conservation", "Logging", "Hurricane", "Fire", "Location", "Landscape", "Lat", "Long", "Season")
colnames(sitecovs2015) <- c("site", "Brand", "Conservation", "Logging", "Hurricane", "Fire", "Location", "Landscape", "Lat", "Long", "Season")
colnames(sitecovs2016) <- c("site", "Brand", "Conservation", "Logging", "Hurricane", "Fire", "Location", "Landscape", "Lat", "Long", "Season")
colnames(sitecovs2017) <- c("site", "Brand", "Conservation", "Logging", "Hurricane", "Fire", "Location", "Landscape", "Lat", "Long", "Season")
colnames(sitecovs2018) <- c("site", "Brand", "Conservation", "Logging", "Hurricane", "Fire", "Location", "Landscape", "Lat", "Long", "Season")
colnames(sitecovs2019) <- c("site", "Brand", "Conservation", "Logging", "Hurricane", "Fire", "Location", "Landscape", "Lat", "Long", "Season")
colnames(sitecovs2020) <- c("site", "Brand", "Conservation", "Logging", "Hurricane", "Fire", "Location", "Landscape", "Lat", "Long", "Season")

#bind1
sitecovs_tojoin <- rbind(sitecovs2014, sitecovs2015, sitecovs2016, sitecovs2017, sitecovs2018, sitecovs2019, sitecovs2020)
```

```{r continue}
##other site covs
require(readxl)
covs <- read_excel("Covariates_cams.xlsx")
covs <- covs %>%
  select(CameraID, Canopy_Height_m)

#join new variables
biomass <- read.csv("mean_biomass.csv")
forest_loss <- read_excel("FL_new.xlsx")
#dis to roads
#dis to streams
ndvi <- read.csv("NDVI.csv")

#fix FL cov
forest_loss$Forest_Los <- round(forest_loss$FL, digits = 1)
forest_loss <- forest_loss[,-2]
forest_loss$FL <- ifelse(forest_loss$Forest_Los == 0, "No Loss", "Loss")

#join to covs
biomass <- left_join(biomass, forest_loss, by = "CameraID")
biomass <- biomass[!duplicated(biomass$CameraID), ]

#join NDVI
biomass <- left_join(biomass, ndvi, by = c("CameraID" = "Name"))
biomass <- biomass[!duplicated(biomass$CameraID), ]

covs <- left_join(biomass, covs, by = "CameraID")
covs <- covs[!duplicated(covs$CameraID), ]

#join cov frames together
covs_all <- left_join(sitecovs_tojoin, covs, by = c("site" = "CameraID"))

covs_all$LatLong <- (covs_all$Lat * covs_all$Long) #new Lat*Long column to account for autocorrelation

#read in distance to covariates 
disturb <- read.csv("NATDISTURB.csv")
disturb_select <- disturb %>%
  select(site, NEAR_DIST_ROAD, NEAR_DIST_STRM, Nat_Dist)

covs_all <- left_join(covs_all, disturb_select, by = "site")
#save
save(covs_all, file = "covs_all.RData")
write.csv("covariates_all.csv", row.names = FALSE)
```
