==> 03-Analysis_function.Rmd <==
---
title: "Analysis_Function"
author: "Elise Boos"
date: "2022-11-22"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## SET UP 

```{r set up }
#clean space
rm(list=ls())

#load packages
#packages
require(tidyverse)
require(dplyr)
require(sf)
require(leaflet)
require(mapview)
require(readxl)
require(lubridate)
require(readxl)
library(unmarked)
library(AICcmodavg)
library(camtrapR)
library(ubms)


##load dfs
load("dat.RData")
load("camop.RData")
covs_all <- read.csv("covs_fin.csv",stringsAsFactors = TRUE)

```

## FINAL CLEANING

```{r final cleaning}
#fix logging column
#covs_all$Logging <- ifelse(covs_all$Logging == c("No", "No Logging"), "No", "Yes")

library(forcats)
covs_all$Logging_fin <- fct_collapse(covs_all$Logging, No = c("No","No Logging"), Yes = c("Yes","Logging"))

#fix mean bioma
covs_all$mean_bioma <- replace_na(covs_all$mean_bioma, 117.817)

#fix lat long 
covs_all[covs_all$site =="908F",]$Lat <- 17.66851300
covs_all[covs_all$site =="908F",]$Long <- -89.01283000

#remove 2016
camop <- subset(camop, !(site %in% c("601H",  "601I",  "601J",  "602H",  "602I",  "602J",  "602K",  "603H",  "603J",  "603K",  "604H",  "604I","604J",  "605H",  "605I",  "605J",  "605K",  "606J",  "606K",  "607A",  "607C",  "608A",  "608B",  "608C",  "609A",  "609B",  "609C", "610A",  "610B",  "610C",  "611A",  "611B",  "611C",  "612A",  "612C",  "614H",  "614I",  "614J")))
camop$site <- droplevels(camop$site)

covs_all <- subset(covs_all, !(site %in% c("601H",  "601I",  "601J",  "602H",  "602I",  "602J",  "602K",  "603H",  "603J",  "603K",  "604H",  "604I","604J",  "605H",  "605I",  "605J",  "605K",  "606J",  "606K",  "607A",  "607C",  "608A",  "608B",  "608C",  "609A",  "609B",  "609C", "610A",  "610B",  "610C",  "611A",  "611B",  "611C",  "612A",  "612C",  "614H",  "614I",  "614J")))
covs_all$site <- droplevels(covs_all$site)

dat <- subset(dat, !(site %in% c("601H",  "601I",  "601J",  "602H",  "602I",  "602J",  "602K",  "603H",  "603J",  "603K",  "604H",  "604I","604J",  "605H",  "605I",  "605J",  "605K",  "606J",  "606K",  "607A",  "607C",  "608A",  "608B",  "608C",  "609A",  "609B",  "609C", "610A",  "610B",  "610C",  "611A",  "611B",  "611C",  "612A",  "612C",  "614H",  "614I",  "614J")))
dat$site <- droplevels(dat$site)

# #look at biomass vs logging
# plot <- ggplot(covs_all, aes(x = Logging, y = mean_bioma, color = Logging)) +
#   geom_line() +
#   geom_point()
# 
# #look at logging & FL columns
# plot2 <- ggplot(covs_all, aes(x = Logging, y = Forest_Los, color = Logging)) +
#   geom_line() +
#   geom_point()

# xtabs(~Logging + FL, data = covs_all) #noteworthy - we have much more logging sites than no logging sites
# #92 sites no logging, 318 sites yes logging
# apply(table(covs_all[, c("Logging", "FL")]), 1, prop.table)

#make camera_operation frame
cam_op <- cameraOperation(camop, stationCol = "site", setupCol = "Date.Placement", 
                             retrievalCol = "Last.record", occasionStartTime = 0 , 
                             dateFormat = "%Y-%m-%d", writecsv = FALSE)


#detection history for gray fox

#getting an error for date_time_obs so checking it
IsDate <- function(mydate, date.format = "%Y-%m-%d %H.%M") {
  tryCatch(!is.na(as.Date(mydate, date.format)),  
           error = function(err) {FALSE})  
}

datecheck <- data.frame(IsDate(dat$date_time_obs))
#issues with rows
dat <- dat[-c(1557,4485, 11912, 13478),]
```

```{r save cleaned data}
save(camop, file = "./clean_data/camera_data.RData")
save(cam_op, file = "./clean_data/cameras_operation.RData")
save(covs_all, file = "./clean_data/covariates.RData")
save(dat, file = "./clean_data/detections_data.RData")

```


```{r mod list}
#species vector 
#mammals <- dat[dat$Mammal == "Yes",]
#unique(mammals$Species)
#count <- summary(mammals$Species)
#list <- subset(count, count > 10)
#species_list <- names(list)
#print(species_list, quote=F)



count <- summary(dat$Species)
#count
list <- subset(count, count > 40)
#list #15991 obs >40
species_list <- names(list)
species_list
#specify range to run
species_list <- species_list[1:10]
species_list
#JUST GRAY FOX 
#species_list <- "Gray Fox"

#start modeling
modlist <- list()
```

```{r make function}
for (species in species_list){
    detect_hist <- detectionHistory(recordTable = dat, species = species, 
                                       camOp = cam_op, output = "binary", stationCol = "site", 
                                       speciesCol = "Species", day1 ="station", 
                                       recordDateTimeCol = "date_time_obs", 
                                       recordDateTimeFormat =  "%Y-%m-%d %H.%M", 
                                       timeZone = "UTC", occasionLength = 1, 
                                       includeEffort = TRUE, scaleEffort = FALSE, 
                                       writecsv = FALSE)
    
    unmarkedFrame <- unmarkedFrameOccu(y = detect_hist$detection_history, siteCovs = covs_all)
    
    unmarkedFrame@siteCovs$Canopy_Height_m <- scale(unmarkedFrame@siteCovs$Canopy_Height_m)
    unmarkedFrame@siteCovs$mean_bioma <- scale(unmarkedFrame@siteCovs$mean_bioma)
    unmarkedFrame@siteCovs$NEAR_DIST_ROAD <- scale(unmarkedFrame@siteCovs$NEAR_DIST_ROAD)
    unmarkedFrame@siteCovs$NEAR_DIST_STRM <- scale(unmarkedFrame@siteCovs$NEAR_DIST_STRM)
    unmarkedFrame@siteCovs$mean_NDVI <- scale(unmarkedFrame@siteCovs$mean_NDVI)
  
    #UNCOMMENT MODELS TO RUN, FOR NOW JUST DOING FULL MODEL
    
     modlist[[paste0(species,"null")]] <-stan_occu(data = unmarkedFrame, formula = ~1 ~1, chains = 4, iter = 100000)
    # modlist[[paste0(species,"ANTHRO")]] <- stan_occu(data = unmarkedFrame, formula = ~1 ~Logging + NEAR_DIST_ROAD + mean_bioma + (1|site), chains = 4, iter = 100000)
    # modlist[[paste0(species,"ENVR")]] <- stan_occu(data = unmarkedFrame, formula = ~1 ~ NEAR_DIST_STRM + mean_NDVI + Season + Nat_Dist + (1|site), chains = 4, iter = 100000)
    modlist[[paste0(species,"ANTHRO+ENVR")]] <- stan_occu(data = unmarkedFrame, formula = ~1 ~Logging + NEAR_DIST_ROAD + mean_bioma + NEAR_DIST_STRM + mean_NDVI + Season + Nat_Dist + (1|site), chains = 4, iter = 100000)
}

```


```{r examine}
oturkey_null <- modlist$`Ocellated Turkeynull`
save(oturkey_null, file = "oturkey_null.RData")
oturkey_anthro_env <- modlist$`Ocellated TurkeyANTHRO+ENVR`
save(oturkey_anthro_env, file = "oturkey_anthro_env.RData")
gcurassow_null <- modlist$`Great Curassownull`
save(gcurassow_null, file = "gcurassow_null.RData")
gcurassow_anthro_env <- modlist$`Great CurassowANTHRO+ENVR`
save(gcurassow_anthro_env, file = "gcurassow_anthro_env.RData")
gfox_null <- modlist$`Gray Foxnull`
save(gfox_null, file = "gfox_null.RData")
gfox_anthro_env <- modlist$`Gray FoxANTHRO+ENVR`
save(gfox_anthro_env , file = "gfox_anthro_env .RData")
wtdeer_null <- modlist$`White-tailed Deernull`
save(wtdeer_null, file = "wtdeer_null.RData")
wtdeer_anthro_env <- modlist$`White-tailed DeerANTHRO+ENVR`
save(wtdeer_anthro_env, file = "wtdeer_anthro_env.RData")
agouti_null <- modlist$Agoutinull
save(agouti_null, file = "agouti_null.RData")
agouti_anthro_env <- modlist$`AgoutiANTHRO+ENVR`
save(agouti_anthro_env , file = "agouti_anthro_env .RData")
ocelot_null <- modlist$Ocelotnull
save(ocelot_null, file = "ocelot_null.RData")
ocelot_anthro_env <- modlist$`OcelotANTHRO+ENVR`
save(ocelot_anthro_env, file = "ocelot_anthro_env.RData")
puma_null <- modlist$Pumanull
save(puma_null, file = "puma_null.RData")
puma_anthro_env <- modlist$`PumaANTHRO+ENVR`
save(puma_anthro_env, file = "puma_anthro_env.RData")
paca_null <- modlist$Pacanull
save(paca_null, file = "paca_null.RData")
paca_anthro_env <- modlist$`PacaANTHRO+ENVR`
save(paca_anthro_env, file = "paca_anthro_env.RData")
ghdove_null <- modlist$`Gray-headed Dovenull`
save(ghdove_null, file = "ghdove_null.RData")
ghdove_anthro_env <- modlist$`Gray-headed DoveANTHRO+ENVR`
save(ghdove_anthro_env, file = "ghdove_anthro_env.RData")
rbdeer_null <- modlist$`Red Brocket Deernull`
save(rbdeer_null, file = "rbdeer_null.RData")
rbdeer_anthro_env <- modlist$`Red Brocket DeerANTHRO+ENVR`
save(rbdeer_anthro_env, file = "rbdeer_anthro_env.RData")

# grayfoxmodels_2_null <- modlist$`Gray Foxnull`
# grayfoxmodels_2_anthro <- modlist$`Gray FoxANTHRO`
# grayfoxmodels_2_envr <- modlist$`Gray FoxENVR`
# grayfoxmodels_2_anthro_envr <- modlist$`Gray FoxANTHRO+ENVR`
# save(grayfoxmodels_2_null, file = "grayfoxmodels_2_null.RData")
# save(grayfoxmodels_2_anthro, file = "grayfoxmodels_2_anthro.RData")
# save(grayfoxmodels_2_envr, file = "grayfoxmodels_2_envr.RData")
# save(grayfoxmodels_2_anthro_envr, file = "grayfoxmodels_2_anthro_envr.RData")
  # Agoutinull <- modlist$Agoutinull
  # AgoutiANTHRO <- modlist$AgoutiANTHRO
  # AgoutiENVR <- modlist$AgoutiENVR
  # modlist$`Ameiva undatanull`
  # modlist$`Ameiva undataANTHRO`
  # modlist$`Ameiva undataENVR`
  # modlist$`Bare-throated Tiger Heronnull`
  # modlist$`Bare-throated Tiger HeronANTHRO`
  # modlist$`Bare-throated Tiger HeronENVR`
  # 
  # save(covs_all, file = "covs_all.RData")
  # 
  # mammals <- dat[dat$Mammal == "Yes",]
  # unique(mammals$Species)
  # count <- summary(mammals$Species)
  # list <- subset(count, count > 10)
  # species_list <- names(list)

# load("ocelotmodels.RData")
# load("agouti_models.RData")
# load("grayfoxmodels.RData")
# load("whitetaileddeermodels.RData")
# load("pumamodels.RData")

# grayfoxmodels_2_anthro_envr
```

```{r posteriors etc}
# plot_posteriors(grayfoxmodels_2_anthro)
# plot_residuals(grayfoxmodels_2_anthro, "state")
# plot_marginal(grayfoxmodels_2_anthro, "state")
# 
# plot_posteriors(grayfoxmodels_2_envr)
# plot_residuals(grayfoxmodels_2_envr, "state")
# plot_marginal(grayfoxmodels_2_envr, "state")
```
