==> 04-Results.Rmd <==
---
title: "Results"
author: "KB"
date: "2023-03-04"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
#Results

## SET UP 

```{r set up }
#clean space
rm(list=ls())

#load packages
#packages
require(tidyverse)
require(dplyr)
require(sf)
require(leaflet)
require(mapview)
require(readxl)
require(lubridate)
require(readxl)
library(unmarked)
library(AICcmodavg)
library(camtrapR)
library(ubms)
library(gridExtra)

```

##Great Curassow

```{r Great Curasow}
#set directory to location of models
setwd("D:/Model Outputs/Feb21")

##load models
load("greatcura_null.RData")
load("greatcura_env.RData")
load("greatcura_anthro.RData")
load("greatcura_anthro_env.RData")

#fit_greatcura <- cbind(null = coef(greatcura_null), env = coef(greatcura_env), anthro = coef(greatcura_anthro), anthro_env = coef(greatcura_anthro_env))

#save(fit_greatcura, file = "fit_greatCura.Rda")
load("fit_greatcura.Rda")

names(greatcura_anthro_env)

#posterior for intercept
occ_intercept <- extract(greatcura_anthro_env, "beta_state[(Intercept)]")[[1]]
hist(occ_intercept, freq=FALSE)
lines(density(occ_intercept), col='red', lwd=2)
#posterior for logging
occ_log <- extract(greatcura_anthro_env, "beta_state[LoggingYes]")[[1]]
hist(occ_intercept, freq=FALSE)
lines(density(occ_intercept), col='red', lwd=2)
```

```{r model comparison }
#First we combine the models into a fitList
mods_greatcura <- fitList(greatcura_null, greatcura_env, greatcura_anthro, greatcura_anthro_env)
#models are compared using leave-out-one cross-validation (LOO) via the loo package. Expected predictive accuracy (elpd) for each model is calculated & the model with the largest elpd performed best.

round(modSel(mods_greatcura), 3)
#anthropogenic model preformed best, but very close to anthro_env

#remove others
#rm(greatcura_env, greatcura_null)

greatcura_anthro_env
```

```{r diagnostics and model fit}

fit_top <- greatcura_anthro
fit_top
#looking at summary: all chains have adequately converged based on R hat values (ideally R > 1.05)
#effective sample size n_eff looks okay (generally want n_eff > (100*chains so 400))

traceplot(fit_top) #visualize convergence (you can specify parameters here or it will show the first 10)

#calculate residuals against covariate values
plot_posteriors(fit_top)
plot_residuals(fit_top, "state") #if model fit is good you expect 95% of binned residuals to fall within shaded area
residuals(fit_top)

#goodness-of-fit test (posterior predictive checks) M-B chi-square test via gof function. Proportion of draws where simulated stat is larger than actual stat should be near 0.5 if model fits well.
#fit_top_gof <- gof(fit_top, quiet = TRUE) #don't have enough memory apparently?
#plot(fit_top_gof)


```

```{r model inference}

plot_marginal(fit_top, "state") #plot marginal effects of each covariate with others held at mean
#slighty greater psi with logging, 
#both distance to road and mean biomass have a negative effect on occupancy probability, and neither contain zero


```

```{r predictions}
#head(predict(fit_top, submodel = "state")) #if you want predicted values for each site or obs

#supply newdata as a data.frame to predict specific values
nd <- data.frame(Logging = "No", NEAR_DIST_ROAD = 2.75, mean_bioma = 0, site = "403C")
predict(fit_top, submodel = "state", newdata = nd)
```

##Agouti


```{r Agouti}
#clean space
rm(list=ls())

#set directory to location of models
setwd("D:/Model Outputs/Feb21")

##load models
load("agouti_null.RData")
#load("agouti_env.RData") missing
load("agouti_anthro.RData")
load("agouti_anthro_env.RData")
```

```{r model comparison }
#First we combine the models into a fitList
mods_agouti <- fitList(agouti_null, agouti_anthro, agouti_anthro_env)
#models are compared using leave-out-one cross-validation (LOO) via the loo package. Expected predictive accuracy (elpd) for each model is calculated & the model with the largest elpd performed best.

round(modSel(mods_agouti), 3)
#anthro_env model preformed best

```

```{r diagnostics and model fit}

fit_top <- agouti_anthro_env
fit_top
#looking at summary: all chains have adequately converged based on R hat values (ideally R > 1.05)
#effective sample size n_eff looks okay (generally want n_eff > (100*chains so 400))




#calculate residuals against covariate values
#plot_posteriors(fit_top)
plot_residuals(fit_top, "state") #if model fit is good you expect 95% of binned residuals to fall within shaded area

#goodness-of-fit test (posterior predictive checks) M-B chi-square test via gof function. Proportion of draws where simulated stat is larger than actual stat should be near 0.5 if model fits well.
#fit_top_gof <- gof(fit_top, quiet = TRUE) #don't have enough memory apparently?
#plot(fit_top_gof)


```

```{r model inference}

plot_marginal(fit_top, "state") #plot marginal effects of each covariate with others held at mean
#slighty greater occupancy prob with logging, 
#much higher occupancy prob without nat disturbance

```

```{r predictions}
#head(predict(fit_top, submodel = "state")) #if you want predicted values for each site or obs

#supply newdata as a data.frame to predict specific values
# nd <- data.frame(Logging = "No", NEAR_DIST_ROAD = 2.75, mean_bioma = 0, site = "403C")
# predict(fit_top, submodel = "state", newdata = nd)
```

##Gray Fox


```{r GF}
#clean space
rm(list=ls())

#set directory to location of models
setwd("D:/Model Outputs/Feb21")
load("grayfox_anthro_env.RData")
##load models
#load("grayfox_null.RData")
#load("grayfox_env.RData")
load("grayfox_anthro.RData")


grayfox_anthro_env
```

```{r model comparison }
#First we combine the models into a fitList
mods_grayfox <- fitList(grayfox_null, grayfox_env, grayfox_anthro, grayfox_anthro_env)
#models are compared using leave-out-one cross-validation (LOO) via the loo package. Expected predictive accuracy (elpd) for each model is calculated & the model with the largest elpd performed best.

round(modSel(mods_grayfox), 3)
#anthro_env model preformed best

```

```{r diagnostics and model fit}

fit_top <- grayfox_anthro_env
fit_top
#looking at summary: all chains have adequately converged based on R hat values (ideally R > 1.05)
#effective sample size n_eff looks okay (generally want n_eff > (100*chains so 400))

#calculate residuals against covariate values
#plot_posteriors(fit_top)
plot_residuals(fit_top, "state") #if model fit is good you expect 95% of binned residuals to fall within shaded area

#goodness-of-fit test (posterior predictive checks) M-B chi-square test via gof function. Proportion of draws where simulated stat is larger than actual stat should be near 0.5 if model fits well.
#fit_top_gof <- gof(fit_top, quiet = TRUE) #don't have enough memory apparently?
#plot(fit_top_gof)


```

```{r model inference}

plot_marginal(fit_top, "state") #plot marginal effects of each covariate with others held at mean
#slighty greater occupancy prob with logging, 
#much higher occupancy prob without nat disturbance

```

```{r predictions}
#head(predict(fit_top, submodel = "state")) #if you want predicted values for each site or obs

#supply newdata as a data.frame to predict specific values
# nd <- data.frame(Logging = "No", NEAR_DIST_ROAD = 2.75, mean_bioma = 0, site = "403C")
# predict(fit_top, submodel = "state", newdata = nd)
```

##Baird's Tapir


```{r tapir}
#clean space
rm(list=ls())

#set directory to location of models
setwd("E:/Model Outputs")
#load models
load("Tapir_anthro_envr.RData")

#model summary
Tapir_anthro_env
```

```{r model comparison }
#First we combine the models into a fitList
mods_tapir <- fitList(Tapir_null, Tapir_envr, Tapir_anthro, Tapir_anthro_env)
#models are compared using leave-out-one cross-validation (LOO) via the loo package. Expected predictive accuracy (elpd) for each model is calculated & the model with the largest elpd performed best.

round(modSel(mods_tapir), 3)
#anthro_env model preformed best

```

```{r diagnostics and model fit}

fit_top <- Tapir_anthro_env
fit_top
#looking at summary: all chains have adequately converged based on R hat values (ideally R > 1.05)
#effective sample size n_eff looks okay (generally want n_eff > (100*chains so 400))

#calculate residuals against covariate values
#plot_posteriors(fit_top)
plot_residuals(fit_top, "state") #if model fit is good you expect 95% of binned residuals to fall within shaded area

#goodness-of-fit test (posterior predictive checks) M-B chi-square test via gof function. Proportion of draws where simulated stat is larger than actual stat should be near 0.5 if model fits well.
#fit_top_gof <- gof(fit_top, quiet = TRUE) #don't have enough memory apparently?
#plot(fit_top_gof)


```

```{r model inference}

plot_marginal(fit_top, "state") #plot marginal effects of each covariate with others held at mean

```

```{r predictions}
#head(predict(fit_top, submodel = "state")) #if you want predicted values for each site or obs

#supply newdata as a data.frame to predict specific values
# nd <- data.frame(Logging = "No", NEAR_DIST_ROAD = 2.75, mean_bioma = 0, site = "403C")
# predict(fit_top, submodel = "state", newdata = nd)
```

##Effect sizes

```{r modlists}
#set directory to location of models


##load models
setwd("F:/Model Outputs/April17")
# load("agouti_anthro_env .RData")
# load("btapir_anthro_env.RData")
# load("cguan_anthro_env.RData")
load("copossum_anthro_env.RData")
# # load("ghdove_anthro_env.RData")
# # load("gcurassow_anthro_env.RData")
# # load("gfox_anthro_env .RData")
# load("jaguar_anthro_env.RData")
# # load("oturkey_anthro_env.RData")
# # load("ocelot_anthro_env.RData")
# # load("paca_anthro_env.RData")
# # load("puma_anthro_env.RData")
# # load("rbdeer_anthro_env.RData")
# load("wncoati_anthro_env.RData")
# # load("wtdeer_anthro_env.RData")
# # load("collaredpeccary_anthro_env.RData")
# load("gtinamou_anthro_env.RData")
# # load("margay_anthro_env.RData")
# load("rqdove_anthro_env.RData")
# load("stripedskunk_anthro_env.RData")
# load("tayra_anthro_env.RData")
# load("tigerheron_anthro_env.RData")
# load("wlippedpeccary_anthro_env.RData")
##add other possum

#models_list will be a list of anthro_env models for the species
#species with be the name of the model 

#intercept
modlist2 <- list()

modlist2 <- c(copossum_anthro_env)

names(modlist2) <- c("Common Opossum")

species <- names(modlist2)


df_int <- data.frame(species)
df_int$effect <- NA
df_int$`2.5` <- NA
df_int$`97.5` <- NA
df_int$variable <- "Intercept"

for (model in species){
  fit_top <- modlist2[[model]]
  sum <- summary(fit_top, "state")
  coef <- sum$mean[1]
  ci2 <- sum$`2.5%`[1]
  ci9 <- sum$`97.5%`[1]
  df_int[df_int$species == model,]$effect <- coef
  df_int[df_int$species == model,]$`2.5` <- ci2
  df_int[df_int$species == model,]$`97.5` <- ci9
}

#save(df_int, file = "df_int.Rda")
#write.csv(df_int, file = "df_int.csv")


#logging

df_log <- data.frame(species)
df_log$effect <- NA
df_log$`2.5` <- NA
df_log$`97.5` <- NA
df_log$variable <- "Logging"


for (model in species){
  fit_top <- modlist2[[model]]
  sum <- summary(fit_top, "state")
  coef <- sum$mean[2]
  ci2 <- sum$`2.5%`[2]
  ci9 <-sum$`97.5%`[2]
  df_log[df_log$species == model,]$effect <- coef
  df_log[df_log$species == model,]$`2.5` <- ci2
  df_log[df_log$species == model,]$`97.5` <- ci9
}

save(df_log, file = 'df_logging2.Rda')

#Distance to Roads

df_roads <- data.frame(species)
df_roads$effect <- NA
df_roads$`2.5` <- NA
df_roads$`97.5` <- NA
df_roads$variable <- "Distance to Roads"


for (model in species){
  fit_top <- modlist2[[model]]
  sum <- summary(fit_top, "state")
  coef <- sum$mean[3]
  ci2 <- sum$`2.5%`[3]
  ci9 <-sum$`97.5%`[3]
 df_roads[df_roads$species == model,]$effect <- coef
  df_roads[df_roads$species == model,]$`2.5` <- ci2
  df_roads[df_roads$species == model,]$`97.5` <- ci9
}

#Mean Biomass

df_bio <- data.frame(species)
df_bio$effect <- NA
df_bio$`2.5` <- NA
df_bio$`97.5` <- NA
df_bio$variable <- "Mean Biomass"


for (model in species){
  fit_top <- modlist2[[model]]
  sum <- summary(fit_top, "state")
  coef <- sum$mean[4]
  ci2 <- sum$`2.5%`[4]
  ci9 <-sum$`97.5%`[4]
  df_bio[df_bio$species == model,]$effect <- coef
  df_bio[df_bio$species == model,]$`2.5` <- ci2
  df_bio[df_bio$species == model,]$`97.5` <- ci9
}

#Distance to Streams

df_streams <- data.frame(species)
df_streams$effect <- NA
df_streams$`2.5` <- NA
df_streams$`97.5` <- NA
df_streams$variable <- "Distance to Streams"


for (model in species){
  fit_top <- modlist2[[model]]
  sum <- summary(fit_top, "state")
  coef <- sum$mean[5]
  ci2 <- sum$`2.5%`[5]
  ci9 <-sum$`97.5%`[5]
  df_streams[df_streams$species == model,]$effect <- coef
  df_streams[df_streams$species == model,]$`2.5` <- ci2
  df_streams[df_streams$species == model,]$`97.5` <- ci9
}

#NDVI

df_NDVI <- data.frame(species)
df_NDVI$effect <- NA
df_NDVI$`2.5` <- NA
df_NDVI$`97.5` <- NA
df_NDVI$variable <- "NDVI"


for (model in species){
  fit_top <- modlist2[[model]]
  sum <- summary(fit_top, "state")
  coef <- sum$mean[6]
  ci2 <- sum$`2.5%`[6]
  ci9 <-sum$`97.5%`[6]
  df_NDVI[df_NDVI$species == model,]$effect <- coef
  df_NDVI[df_NDVI$species == model,]$`2.5` <- ci2
  df_NDVI[df_NDVI$species == model,]$`97.5` <- ci9
}

#Season (wet)

df_season <- data.frame(species)
df_season$effect <- NA
df_season$`2.5` <- NA
df_season$`97.5` <- NA
df_season$variable <- "Season"


for (model in species){
  fit_top <- modlist2[[model]]
  sum <- summary(fit_top, "state")
  coef <- sum$mean[7]
  ci2 <- sum$`2.5%`[7]
  ci9 <-sum$`97.5%`[7]
  df_season[df_season$species == model,]$effect <- coef
  df_season[df_season$species == model,]$`2.5` <- ci2
  df_season[df_season$species == model,]$`97.5` <- ci9
}

#Natural Disturbance

df_dist <- data.frame(species)
df_dist$effect <- NA
df_dist$`2.5` <- NA
df_dist$`97.5` <- NA
df_dist$variable <- "Natural Disturbance"


for (model in species){
  fit_top <- modlist2[[model]]
  sum <- summary(fit_top, "state")
  coef <- sum$mean[8]
  ci2 <- sum$`2.5%`[8]
  ci9 <-sum$`97.5%`[8]
  df_dist[df_dist$species == model,]$effect <- coef
  df_dist[df_dist$species == model,]$`2.5` <- ci2
  df_dist[df_dist$species == model,]$`97.5` <- ci9
}

#Site

df_site <- data.frame(species)
df_site$effect <- NA
df_site$`2.5` <- NA
df_site$`97.5` <- NA
df_site$variable <- "Site"


for (model in species){
  fit_top <- modlist2[[model]]
  sum <- summary(fit_top, "state")
  coef <- sum$mean[9]
  ci2 <- sum$`2.5%`[9]
  ci9 <-sum$`97.5%`[9]
  df_site[df_site$species == model,]$effect <- coef
  df_site[df_site$species == model,]$`2.5` <- ci2
  df_site[df_site$species == model,]$`97.5` <- ci9
}

CI_copossum <- rbind(df_int, df_log, df_roads, df_bio, df_streams, df_NDVI, df_season, df_dist, df_site)

save(CI_copossum, file = "effects_new_copossum.Rda")


```

##Plot effects

```{r Plotting}

#effects plot
load("effects_new1.Rda")
load("effects_new2.Rda")
load("effects_new3.Rda")
load("effects_new4.Rda")
load("effects_new_margay.Rda")
load("effects_new_copossum.Rda")


CI_new1$variable <- as.factor(CI_new1$variable)
CI_new2$variable <- as.factor(CI_new2$variable)
CI_new3$variable <- as.factor(CI_new3$variable)
CI_new4$variable <- as.factor(CI_new4$variable)
CI_new_margay$variable <- as.factor(CI_new_margay$variable)
CI_copossum$variable <- as.factor(CI_copossum$variable)

CI_new1$species <- as.factor(CI_new1$species)
CI_new2$species <- as.factor(CI_new2$species)
CI_new3$species <- as.factor(CI_new3$species)
CI_new4$species <- as.factor(CI_new4$species)
CI_new_margay$species <- as.factor(CI_new_margay$species)
CI_copossum$species <- as.factor(CI_copossum$species)

CI <- rbind(CI_new1, CI_new2, CI_new3, CI_new4, CI_new_margay, CI_copossum)
CI$species <- recode_factor(CI$species, "Great curassow" = "Great curassow*",
                            "Jaguar" = "Jaguar*",
                            "Ocellated turkey" = "Ocellated turkey*",
                            "Ocelot" = "Ocelot*",
                            "Puma" = "Puma*",
                            "Red brocket" = "Red brocket*",
                            "Margay" = "Margay*",
                            "Baird's tapir" = "Baird's tapir*",
                            "Russet-naped woodrail" = "Russet-naped wood rail",
                            "Crested Guan" = "Crested guan",
                            "Gray-headed Dove" = "Gray-headed dove",
                            "Gray Fox" = "Gray fox",
                            "White-nosed Coati" = "White-nosed coati",
                            "White-tailed Deer" = "White-tailed deer",
                            "Common Opossum" = "Common opossum")

colnames(CI)[1] <- "Species"
colnames(CI)[2] <- "Effect Size"
colnames(CI)[3] <- "X2.5"
colnames(CI)[4] <- "X97.5"
colnames(CI)[5] <- "Cov"

CI$Species <- as.factor(CI$Species)
CI$Cov <- as.factor(CI$Cov)


#does the value cross zero?
CI$pass <- ifelse(CI$X2.5 < 0  & CI$X97.5 >= 0 | CI$X2.5 > 0  & CI$X97.5 <= 0, 0, 1)
#0 means passes
CI$pass <- as.factor(CI$pass)

#Logging
Logging <- CI %>% group_by(Species, Cov) %>% filter(Cov == "Logging")

Logging$Species <- factor(Logging$Species, levels = c("Russet-naped wood rail","Red brocket*","White-tailed deer","White-lipped peccary","Collared peccary","Gray fox","Jaguar*","Ocelot*", "Puma*","Margay*","Striped hog-nosed skunk","Tayra","White-nosed coati","Nine-banded armadillo","Gray-headed dove","Ruddy quail-dove","Common opossum","Brown four-eyed opossum","Crested guan", "Great curassow*", "Ocellated turkey*","Wood thrush","Bare-throated tiger heron","Baird's tapir*","Paca","Agouti","Great tinamou"))

Logging_plot <- ggplot(
  Logging, 
  aes(x = `Effect Size`, y = Species, xmin = X2.5, xmax =X97.5)) +
  geom_point(fill = Logging$pass,
        shape = 21, 
        size = 4) +
  geom_errorbarh(aes(height=.2)) +
  theme(legend.position = "none") +
  theme(axis.title.y=element_blank()) +
  geom_vline(xintercept = 0, linetype="dashed", color = "tomato1") +
  theme_bw() +
  scale_y_discrete(limits = rev)
Logging_plot <- Logging_plot + labs(title = "Logging", size = 15) + theme(axis.title = element_text(size= 15))

ggsave("Logging_plot_june.png", units = "in", width = 8, height = 9)


#Distance to Roads
Dist_Roads <- CI %>% group_by(Species, Cov) %>% filter(Cov == "Distance to Roads")

Dist_Roads$Species <- factor(Dist_Roads$Species, levels = c("Russet-naped wood rail","Red brocket*","White-tailed deer","White-lipped peccary","Collared peccary","Gray fox","Jaguar*","Ocelot*", "Puma*","Margay*","Striped hog-nosed skunk","Tayra","White-nosed coati","Nine-banded armadillo","Gray-headed dove","Ruddy quail-dove","Common opossum","Brown four-eyed opossum","Crested guan", "Great curassow*", "Ocellated turkey*","Wood thrush","Bare-throated tiger heron","Baird's tapir*","Paca","Agouti","Great tinamou"))

Roads_plot <- ggplot(
  Dist_Roads, 
  aes(x = `Effect Size`, y = Species, xmin = X2.5, xmax =X97.5)) +
  geom_point(fill = Dist_Roads$pass,
        shape = 21, 
        size = 3) +
  geom_errorbarh(aes(height=.2)) + #colour = Species for fill color
  theme_bw() +
  theme(legend.position = "none") +
  theme(axis.title.y=element_blank())+
  geom_vline(xintercept = 0, linetype="dashed")
Roads_plot <- Roads_plot + labs(title = "Roads")

#ggsave("Roads_plot_bw.png")

#Biomass
Biomass <- CI %>% group_by(Species, Cov) %>% filter(Cov == "Mean Biomass")

Biomass$Species <- factor(Biomass$Species, levels = c("Russet-naped wood rail","Red brocket*","White-tailed deer","White-lipped peccary","Collared peccary","Gray fox","Jaguar*","Ocelot*", "Puma*","Margay*","Striped hog-nosed skunk","Tayra","White-nosed coati","Nine-banded armadillo","Gray-headed dove","Ruddy quail-dove","Common opossum","Brown four-eyed opossum","Crested guan", "Great curassow*", "Ocellated turkey*","Wood thrush","Bare-throated tiger heron","Baird's tapir*","Paca","Agouti","Great tinamou"))

Biomass_plot <- ggplot(
  Biomass, 
  aes(x = `Effect Size`, y = Species, xmin = X2.5, xmax =X97.5)) +
  geom_point(fill = Biomass$pass,
        shape = 21, 
        size = 3) +
  geom_errorbarh(aes(height=.2)) +
  theme_bw() +
  theme(legend.position = "none") +
  theme(axis.title.y=element_blank()) +
  geom_vline(xintercept = 0, linetype="dashed")
Biomass_plot <- Biomass_plot + labs(title = "Biomass")

#ggsave("Biomass_plot_bw.png")

#Streams
Streams <- CI %>% group_by(Species, Cov) %>% filter(Cov == "Distance to Streams")

Streams$Species <- factor(Streams$Species, levels = c("Russet-naped wood rail","Red brocket*","White-tailed deer","White-lipped peccary","Collared peccary","Gray fox","Jaguar*","Ocelot*", "Puma*","Margay*","Striped hog-nosed skunk","Tayra","White-nosed coati","Nine-banded armadillo","Gray-headed dove","Ruddy quail-dove","Common opossum","Brown four-eyed opossum","Crested guan", "Great curassow*", "Ocellated turkey*","Wood thrush","Bare-throated tiger heron","Baird's tapir*","Paca","Agouti","Great tinamou"))

Streams_plot <- ggplot(
  Streams, 
  aes(x = `Effect Size`, y = Species, xmin = X2.5, xmax =X97.5)) +
  geom_point(fill = Streams$pass,
        shape = 21, 
        size = 3) +
  geom_errorbarh(aes(height=.2)) +
  theme_bw() +
  theme(legend.position = "none") +
  theme(axis.title.y=element_blank())+
  geom_vline(xintercept = 0, linetype="dashed")
Streams_plot <- Streams_plot + labs(title = "Streams")

#ggsave("Streams_plot_bw.png")

##NDVI
NDVI <- CI %>% group_by(Species, Cov) %>% filter(Cov == "NDVI")

NDVI$Species <- factor(NDVI$Species, levels = c("Russet-naped wood rail","Red brocket*","White-tailed deer","White-lipped peccary","Collared peccary","Gray fox","Jaguar*","Ocelot*", "Puma*","Margay*","Striped hog-nosed skunk","Tayra","White-nosed coati","Nine-banded armadillo","Gray-headed dove","Ruddy quail-dove","Common opossum","Brown four-eyed opossum","Crested guan", "Great curassow*", "Ocellated turkey*","Wood thrush","Bare-throated tiger heron","Baird's tapir*","Paca","Agouti","Great tinamou"))

NDVI_plot <- ggplot(
  NDVI, 
  aes(x = `Effect Size`, y = Species, xmin = X2.5, xmax =X97.5)) +
  geom_point(fill = NDVI$pass,
        shape = 21, 
        size = 3) +
  geom_errorbarh(aes(height=.2)) +
  theme_bw() +
  theme(legend.position = "none") +
  theme(axis.title.y=element_blank())+
  geom_vline(xintercept = 0, linetype="dashed")
NDVI_plot <- NDVI_plot + labs(title = "NDVI")

#ggsave("NDVI_plot_bw.png")

##Season
Season <- CI %>% group_by(Species, Cov) %>% filter(Cov == "Season")

Season$Species <- factor(Season$Species, levels = c("Russet-naped wood rail","Red brocket*","White-tailed deer","White-lipped peccary","Collared peccary","Gray fox","Jaguar*","Ocelot*", "Puma*","Margay*","Striped hog-nosed skunk","Tayra","White-nosed coati","Nine-banded armadillo","Gray-headed dove","Ruddy quail-dove","Common opossum","Brown four-eyed opossum","Crested guan", "Great curassow*", "Ocellated turkey*","Wood thrush","Bare-throated tiger heron","Baird's tapir*","Paca","Agouti","Great tinamou"))

Season_plot <- ggplot(
  Season, 
  aes(x = `Effect Size`, y = Species, xmin = X2.5, xmax =X97.5)) +
  geom_point(fill = Season$pass,
        shape = 21, 
        size = 3) +
  geom_errorbarh(aes(height=.2)) +
  theme_bw() +
  theme(legend.position = "none") +
  theme(axis.title.y=element_blank())+
  geom_vline(xintercept = 0, linetype="dashed")
Season_plot <- Season_plot + labs(title = "Season")

#ggsave("Season_plot_bw.png")

##Natural Disturbance
NatD <- CI %>% group_by(Species, Cov) %>% filter(Cov == "Natural Disturbance")

NatD$Species <- factor(NatD$Species, levels = c("Russet-naped wood rail","Red brocket*","White-tailed deer","White-lipped peccary","Collared peccary","Gray fox","Jaguar*","Ocelot*", "Puma*","Margay*","Striped hog-nosed skunk","Tayra","White-nosed coati","Nine-banded armadillo","Gray-headed dove","Ruddy quail-dove","Common opossum","Brown four-eyed opossum","Crested guan", "Great curassow*", "Ocellated turkey*","Wood thrush","Bare-throated tiger heron","Baird's tapir*","Paca","Agouti","Great tinamou"))

NatD_plot <- ggplot(
  NatD, 
  aes(x = `Effect Size`, y = Species, xmin = X2.5, xmax =X97.5)) +
  geom_point(fill = NatD$pass,
        shape = 21, 
        size = 3) +
  geom_errorbarh(aes(height=.2)) + 
  theme_bw() +
  theme(legend.position = "bottom") +
  theme(axis.title.y=element_blank())+
  geom_vline(xintercept = 0, linetype="dashed")
NatD_plot <- NatD_plot + labs(title = "Natural Disturbance")

#ggsave("Dist_plot_bw.png")

# Create plot with legend
require("gridExtra")
ggp1_legend <- ggplot(NatD, 
  aes(x = `Effect Size`, y = Species, xmin = X2.5, xmax =X97.5)) +
  geom_point(fill = NatD$pass,
        shape = 21, 
        size = 3) +
  geom_errorbarh(aes(height=.2, colour = Species)) +
  theme_bw() +
  theme(axis.title.y=element_blank()) +
  geom_vline(xintercept = 0, linetype="dashed") +
  theme(legend.position = "bottom") +
  scale_y_discrete(limits = rev)
  
# Create user-defined function, which extracts legends from ggplots
extract_legend <- function(my_ggp) {
  step1 <- ggplot_gtable(ggplot_build(my_ggp))
  step2 <- which(sapply(step1$grobs, function(x) x$name) == "guide-box")
  step3 <- step1$grobs[[step2]]
  return(step3)
}

# Apply user-defined function to extract legend
shared_legend <- extract_legend(ggp1_legend)

##grid of plots
#install.packages("gridExtra")
tiff("effectsplotbw_june.tiff", units = "in", width = 10, height = 14, res = 200)
p <- grid.arrange(arrangeGrob(Roads_plot, Biomass_plot, Streams_plot, NDVI_plot, Season_plot, NatD_plot, nrow = 3), ncol = 1, heights = c(14, 1))
dev.off()

```

##Corr Test

```{r correlation test}

load("~/Belize-MP-Bruno-Boos/df_psi3.Rda")
#load("~/Belize-MP-Bruno-Boos/df_psi_secondspecies.Rda")
load("~/Belize-MP-Bruno-Boos/df_psi.Rda")

# colnames(df_psi)[1] <- "psi"
# colnames(df_psi)[2] <- "se"

df_psi4 <- data.frame(psi = c(0.51, 0.80, 0.52, 0.67,	0.70, 0.37, 0.71, 0.32, 0.44, 0.69), se = c(0.015, 0.008, 0.013, 0.011, 0.011, 0.013, 0.009, 0.009, 0.011, 0.01), Speciesname = c("Agouti", "Great curassow", "Jaguar", "Ocellated turkey", "Ocelot", "Paca", "Puma", "Red brocket", "White-nosed coati", "White-tailed deer"))

psi <- rbind(df_psi2, df_psi3, df_psi4)
psi$Speciesname <- as.factor(psi$Speciesname)

#p
load("~/Belize-MP-Bruno-Boos/detprobs.Rda")
load("~/Belize-MP-Bruno-Boos/detprobs2.Rda")
load("~/Belize-MP-Bruno-Boos/detprobs3.Rda")

p <- rbind(df_p, df_p2, df_p3)
p$species <- as.factor(p$species)

psi$psi <- as.numeric(psi$psi)

p$species <- recode_factor(p$species, "Gray-headed Dove" = "Gray-headed dove",
                           "Gray Fox" = "Gray fox",
                           "Great Curassow" = "Great curassow",
                           "Ocellated Turkey" = "Ocellated turkey",
                           "Red Brocket Deer" = "Red brocket",
                           "White-nosed Coati" = "White-nosed coati",
                           "White-tailed Deer" = "White-tailed deer")

df <- left_join(p, psi, by = c("species" = "Speciesname"))

x <- df$mean
y <- df$psi

cor_results <- cor.test(x, y, method = "pearson")

save(p, file = "detection_probs.Rda")
save(psi, file = "occupancy_probs.Rda")

#close (0.078), but fail to reject null - true correlation is equal to 0
```

```{r formatting}

setwd("C:/Users/Kelly/Documents/Belize-MP-Bruno-Boos")
load("dat.RData")
img <- dat
load("camop.RData")
dep <- camop

#using lubridate to format dates
library(lubridate)
#covert date columns from character strings to date objects
# start dates
dep$start_date <- ymd(dep$Date.Placement)

# end dates
dep$end_date   <- ymd(dep$Last.record)

#create column called days and calculate the interval for deployments
dep$days <- interval(dep$start_date, dep$end_date)/ddays(1)
#minimum of 1 days, mean of 44.75 days, max of 143 days

sum(dep$days) #18349 total so obs not correct, figure out how to calculate total deployment days

summary(dep$days) #check for 0s (usually means camera malfunction), NAs and negative numbers (entry issue) We don't detect any issues

#Now we need to convert the img timestamp column
img$date_time_obs <- ymd_hm(img$date_time_obs)

range(img$date_time_obs) #check range 3/18/14 (04:28:00) - 11/28/20 (04:45:00)

table(is.na(img$date_time_obs)) #check for NAs
#6 NAs need to be looked at 1557, 4439, 4743, 6066, 13493, 15059
img <- img[-c(1557,4439,4743,6066,13493,15059),]

```

```{r basic summary}

# Count the number of camera locations
paste(length(unique(dep$site)), "locations"); paste(length(unique(dep$deployment_id)), "deployments");paste(nrow(img), "image labels"); paste(nrow(img[img$is_blank == TRUE,]), "blanks")

#410 locations, 16,734 observations

```
13 locations, 37 deployments, 984 image labels, 0 blanks
