==> 05-Post-hoc.Rmd <==
---
title: "ANCOVA"
author: "Kelly Bruno"
date: "2023-04-03"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## SET UP 

```{r set up }
#clean space
rm(list=ls())

#load packages
#packages
require(tidyverse)
require(dplyr)
require(ggpubr)
require(rstatix)
require(broom)
require(datarium)
```


## Load Files

```{r files}

load("~/Belize-MP-Bruno-Boos/Effects/effects2.Rda")
load("~/Belize-MP-Bruno-Boos/Effects/effects.Rda")
load("~/Belize-MP-Bruno-Boos/Effects/effects_tapir.Rda")

#remove old results
CI <- subset(CI, species != "Baird's tapir")

#bind effects
effect_sizes <- rbind(CI, CI2, CI_tap)

effect_sizes$species <- as.factor(effect_sizes$species)
```

#2-way ANCOVA prep

ANCOVA assumes: 
- linearity between covariate and outcome variable
- homogeneity of regression slopes formed by covariate and outcome variable
- outcome variable approx. normally distributed
- homoscedasticity of residuals variance
- no significant outliers

```{r ANCOVA prep logging}

effect_logging <- effect_sizes %>% filter(variable == "Logging")

#create groups
effect_logging <- effect_logging %>% mutate(class = ifelse(species %in% c("	
Crested Guan", "Gray-headed Dove", "Great Curassow", "Ocellated Turkey", "Crested Guan"), "Aves", "Mammalia"))

#linearity assumption: scatter plot between covariate (logging) and outcome variable (effect size) for each combination of groups


#these should probably be logging yes or logging no with different effect sizes....need to rework
ggscatter(
  effect_logging, x = "variable", y = "effect",
  facet.by  = c("class", "logging_YorN"), 
  short.panel.labs = FALSE
  )+
  stat_smooth(method = "loess", span = 0.9)


```
