==> 07-Probabilities.Rmd <==
---
title: "07-Probabilities"
author: "Kelly Bruno"
date: "2023-04-17"
output: html_document
---

---
title: "Analysis_Function"
author: "Elise Boos"
date: "2022-11-22"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## SET UP 

```{r set up }
#clean space
rm(list=ls())

#load packages
#packages
require(tidyverse)
require(dplyr)
require(sf)
require(leaflet)
require(mapview)
require(readxl)
require(lubridate)
require(readxl)
library(unmarked)
library(AICcmodavg)
library(camtrapR)
library(ubms)

```

## Predict test

```{r predict}
setwd("E:/Model Outputs")
load("agouti_anthro_env.RData")
load("agouti_null.RData")

#test
#posterior_predict(agouti_anthro_env, param = "y", draws = 1, re.form = NULL)

predict(
  agouti_null,
  "state", #submodel
  newdata = NULL,
  transform = TRUE, #back-transform the predictions to their original scale
  level = 0.95)

##occupancy probability
##ask john about this method
Gdove_fit <- data.frame(x = predict(GHDove_anthro_env, "state", newdata = NULL, transform = TRUE, re.form = NULL, level = 0.95))

#summarize prediction data
p_Gdove <- Gdove_fit %>% summarise(psi = mean(x.Predicted), se = sd(x.Predicted, na.rm = TRUE)/sqrt(n())) %>% as.data.frame()

save(p_Gdove, file = "p_Gdove.Rda")

```


```{r loop to predict occ}
#set directory to location of models
#setwd("E:/Model Outputs")

##load models
setwd("D:/Model Outputs/April 3")
# load("agouti_anthro_env.RData")
# load("btapir_anthro_env.RData")
# load("cguan_anthro_env.RData")
# #load("Common_Opossum_anthro_env.RData")
# load("GHDove_anthro_env.RData")
# load("grayfox_anthro_env.RData")
# load("greatcura_anthro_env.RData")
# load("jaguar_anthro_env.RData")
# load("ocellatedturkey_anthro_env.RData")
# load("ocelot_anthro_env.RData")
# load("Paca_anthro_env.RData")
# load("puma_anthro_env.RData")
# load("redbrockdeer_anthro_env.RData")
# load("wnosedcoati_anthro_env.RData")
# load("wtdeer_anthro_env.RData")
# load("BfourOpossum_anthro_env.RData")
# load("CommonOpossum_anthro_env.RData")
# load("NbArmadillo_anthro_env.RData")
# load("RnWoodrail_anthro_env.RData")
# load("Wthrush_anthro_env.RData")

#predlist will be a list of anthro_env models for the species
#species with be the name of the model

#prediction list
predlist <- list()
predlist <- c(BfourOpossum_anthro_env, CommonOpossum_anthro_env, NbArmadillo_anthro_env, RnWoodrail_anthro_env, Wthrush_anthro_env) #agouti_anthro_env, btapir_anthro_env, cguan_anthro_env, GHDove_anthro_env, grayfox_anthro_env, greatcura_anthro_env, jaguar_anthro_env, ocellatedturkey_anthro_env, ocelot_anthro_env, Paca_anthro_env, puma_anthro_env, redbrockdeer_anthro_env, wnosedcoati_anthro_env, wtdeer_anthro_env)

names(predlist) <- c("Brown four-eyed opossum", "Common opossum", "Nine-banded armadillo", "Russet-naped woodrail", "Wood thrush") #"Agouti", "Baird's tapir", "Crested Guan", "Gray-headed Dove", "Gray Fox","Great Curassow", "Jaguar", "Ocellated Turkey", "Ocelot", "Paca", "Puma", "Red Brocket Deer", "White-nosed Coati", "White-tailed Deer")

species <- names(predlist)

##loop to predict occupancy

df_psi <- data.frame(species)
df_psi$mean <- NA
df_psi$se <- NA

for (model in species){
  fit_top <- predlist[[model]]
  df_pred <- data.frame(x = predict(fit_top, "state", newdata = NULL, transform = TRUE, re.form = NULL, level = 0.95))
  df_psi[df_psi$species == model,] <- df_pred %>% summarise(mean = mean(x.Predicted), se = sd(x.Predicted, na.rm = TRUE)/sqrt(n())) %>% as.data.frame()
}

df_psi$Speciesname <- c("Brown four-eyed opossum", "Common opossum", "Nine-banded armadillo", "Russet-naped woodrail", "Wood thrush")
colnames(df_psi) <- c("psi", "se", "NA", "Speciesname")
df_psi <- df_psi[,-3]
df_psi3 <- df_psi

save(df_psi3, file = "df_psi3.Rda")

# p_cguan$Speciesname <- "Crested guan"
# p_fox$Speciesname <- "Gray fox"
# p_Gdove$Speciesname <- "Gray-headed dove"
# p_tapir$Speciesname <- "Baird's tapir"
# 
# df_psi1 <- select(df_psi, "psi" = species,"se" = mean, Speciesname = "Speciesname")
# 
# df_psi2 <- rbind(df_psi1, p_cguan, p_fox, p_Gdove, p_tapir)
# 
# save(df_psi2, file = "df_psi.Rda")
# write.csv(df_psi2, file = "df_psi.csv")

```

```{r loop to predict det}
#set directory to location of models
setwd("E:/Model Outputs/April 3")
require(boot)

##load models
# setwd("D:/Model Outputs")
# load("agouti_anthro_env.RData")
# load("cguan_anthro_env.RData")
# #load("Common_Opossum_anthro_env.RData")
# load("GHDove_anthro_env.RData")
# load("grayfox_anthro_env.RData")
# load("greatcura_anthro_env.RData")
# load("jaguar_anthro_env.RData")
# load("ocellatedturkey_anthro_env.RData")
# load("ocelot_anthro_env.RData")
# load("Paca_anthro_env.RData")
# load("puma_anthro_env.RData")
# load("redbrockdeer_anthro_env.RData")
# load("wnosedcoati_anthro_env.RData")
# load("wtdeer_anthro_env.RData")
# load("BfourOpossum_anthro_env.RData")
# load("CommonOpossum_anthro_env.RData")
# load("NbArmadillo_anthro_env.RData")
# load("RnWoodrail_anthro_env.RData")
# load("Wthrush_anthro_env.RData")
# load("Tapir_anthro_envr.RData")


#modlist will be a list of anthro_env models for the species
#species with be the name of the model

#prediction list
modlist <- list()
modlist <- c(BfourOpossum_anthro_env, CommonOpossum_anthro_env, NbArmadillo_anthro_env, RnWoodrail_anthro_env, Wthrush_anthro_env)
  #c(agouti_anthro_env, cguan_anthro_env, GHDove_anthro_env, grayfox_anthro_env, greatcura_anthro_env, jaguar_anthro_env, ocellatedturkey_anthro_env, ocelot_anthro_env, Paca_anthro_env, puma_anthro_env, redbrockdeer_anthro_env, wnosedcoati_anthro_env, wtdeer_anthro_env)

names(modlist) <- c("Brown four-eyed opossum", "Common opossum", "Nine-banded armadillo", "Russet-naped woodrail", "Wood thrush")
  #c("Agouti", "Crested Guan", "Gray-headed Dove", "Gray Fox","Great Curassow", "Jaguar", "Ocellated Turkey", "Ocelot", "Paca", "Puma", "Red Brocket Deer", "White-nosed Coati", "White-tailed Deer")

species <- names(modlist)

##loop to predict detection

df_p <- data.frame(species)
df_p$mean <- NA
df_p$sd <- NA

for (model in species){
  fit_top <- modlist[[model]]
  sum <- summary(fit_top, "det")
  meanodds <- exp(sum$mean)
  meanprob <- meanodds / (1 + meanodds)
  sdodds <- exp(sum$sd)
  sdprob <- sdodds / (1 + sdodds)
 df_p[df_p$species == model,]$mean <- meanprob
  df_p[df_p$species == model,]$sd <- sdprob
}

df_p2 <- df_p

save(df_p2, file = "detprobs2.Rda")
write.csv(df_p2, file = "detprobs2.csv")

```
