==> 08-Discussion.Rmd <==
---
title: "Model Results Post-Hoc"
author: "Elise Boos"
date: "2023-05-26"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

### MAKING MODEL TO UNDERSTAND RESULTS

```{r effects}
#load packagaes
require(tidyverse)
require(agricolae)

# #load files
# load("effects_new1.Rda")
# load("effects_new2.Rda")
# load("effects_new3.Rda")
# load("effects_new4.Rda")
# load("effects_new_margay.Rda")
# 
# all <- rbind(CI_new_margay, CI_new1, CI_new2, CI_new3, CI_new4)
# save(all, file = "all_effects.Rda")

load("all_effects.RDa")
st <- read.csv("species_traits.csv", stringsAsFactors = T, nrows= 27)

#note - species of concern = belizean SOC or threatened on IUCN list

##DROPPING OPOSSUM FOR NOW 
st2 <- st %>% filter(Common.Name != "Common opossum")

alldat <- full_join(all, st2, by = c("species" ="Common.Name") )

vars <- unique(alldat$variable)
```

```{r}
alldat_bird <- alldat %>%
  filter(Type == "Bird")

alldat_mammal <- alldat %>%
  filter(Type == "Mammal")
install.packages("plotrix")
library(plotrix)
std.error(filter(alldat_bird, variable == "Logging")$effect)
std.error(filter(alldat_mammal, variable == "Logging")$effect)
```

### Model test

```{r linear models}
#test on logging first 

# BY ORDER 

# carn <- alldat %>%
#   filter(Guild == "Carnivore")
# 
# # df_wider <- alldat %>%
# #  # mutate(sp_var = paste(species, variable)) %>%
# #   pivot_wider(names_from = variable, values_from = effect)
# 
# 
# 
#  aov1<- aov(effect ~ variable, data = carn)  #%>%
#  #  filter(variable == "Logging"))
# summary(aov1)
# TukeyHSD(aov1)
# # 
# 
# groups <- HSD.test(aov1, "variable", group = TRUE)
# groups
# 
# 
# # BY MAMMAL OR BIRD 
# 
 aov2 <- aov(effect ~ Type, data = alldat %>%
   filter(variable == "Logging"))
 summary(aov2)
 TukeyHSD(aov2)
# 
 groups <- HSD.test(aov2, "Type", group = TRUE)
 groups
```

###SAVEABLE RESULTS

##### DO NOT NEED TO RUN AGAIN ####

```{r func}
#ANOVA RESULTS
# aovs <- function(var){
#   aov2 <- aov(effect ~ Species_of_concern, data = alldat %>%
#   filter(variable == var))
#   summary(aov2)
# }
# results <- map(vars, aovs)
# names(results) <- vars
# 
# results$Logging
# #pull out p value 
# pvals <- c(results$Intercept[[1]]$`Pr(>F)`[1],results$Logging[[1]]$`Pr(>F)`[1],results$`Distance to Roads`[[1]]$`Pr(>F)`[1], results$`Mean Biomass`[[1]]$`Pr(>F)`[1], results$`Distance to Streams`[[1]]$`Pr(>F)`[1], results$NDVI[[1]]$`Pr(>F)`[1], results$Season[[1]]$`Pr(>F)`[1], results$`Natural Disturbance`[[1]]$`Pr(>F)`[1], results$Site[[1]]$`Pr(>F)`[1])
#            
# pvaldf_soc <- as.data.frame(cbind(vars, pvals))
# pvaldf_soc$pvals <- as.numeric(pvaldf_soc$pvals)

#save groupings 
# save(pvaldf_soc, file = "pvaldf_soc.Rda")
```

```{r}
# p values for each grouping 
load("pvaldf_type.Rda")
load("pvaldf_guild.Rda")
load("pvaldf_soc.Rda")

```


### LOOK AT THE DIFFERENCES IN MEANS BETWEEN THE GORUPINGS

```{r}
# GROUPING RESULTS 

#use this to see the differences in groups 
groups <- function(var){
  aov2 <- aov(effect ~ Species_of_concern, data = alldat %>%
  filter(variable == var))
  HSD.test(aov2, "Species_of_concern", group = TRUE)
}
results2 <- map(vars, groups)
names(results2) <- vars

results2$`Mean Biomass`
results2$Season
results2

pvaldf_type
```

## CORRECT WAY ?

```{r}
# GROUPING RESULTS 
#use this to see the differences in groups 
types <- unique(alldat$Type)

groups <- function(type){
  aov2 <- aov(effect ~ variable, data = alldat %>%
  filter(Type == type))
  HSD.test(aov2, "variable", group = TRUE)
}
results2 <- map(types, groups)
names(results2) <- types

# results2$`Mean Biomass`
# results2$Season
results2

```







#### EXTRA STUFF BELOW NOT NECESSARY 

### USE JUST TO PRINT OUT *ALL* RESULTS FOR EXAMINING 

```{r lm for loop}

# BY ORDER
for (var in vars) {
aov1 <- aov(effect ~ Order, data = alldat %>%
  filter(variable == var))
summary(aov1)
TukeyHSD(aov1)

groups <- HSD.test(aov1, "Order", group = TRUE)

print(var)
print(groups)
}

# BY MAMMAL OR BIRD 
for (var in vars) {
aov2 <- aov(effect ~ Type, data = alldat %>%
  filter(variable == var))

print(var)

print(summary(aov2))
TukeyHSD(aov2)

groups <- HSD.test(aov2, "Type", group = TRUE)

print(groups)
}


# BY GUILD
for (guild in guilds) {
aov2 <- aov(effect ~ variable, data = alldat %>%
  filter(Guild == guild))

print(guild)

print(summary(aov2))
TukeyHSD(aov2)

groups <- HSD.test(aov2, "variable", group = TRUE)

print(groups)
}


# BY SOC
for (var in vars) {
aov2 <- aov(effect ~ Species_of_concern, data = alldat %>%
  filter(variable == var))

print(var)

print(summary(aov2))
TukeyHSD(aov2)

groups <- HSD.test(aov2, "Species_of_concern", group = TRUE)

print(groups)
}
```
